shrink repository traits
Since the actual fossil collection and deletion only use a subset of
the repository traits their implementation is harder than it needs to be.

To reduce the difficulty the traits have been reduced and the file backend
has been moved into the `vinculum-benchmark` crate since it is now an
implementation detail.